refactor: Remove redundant call to updateRestaurants() in window.initMap()
Under specific circumstances, this call would have caused the Restaurants JSON
data to be fetched twice and the render-tree to be updated twice unnecessarily.

The last commit updated the code for performance, with one of the things
improved upon being that we no longer wait for the Google Maps script to be fetched,
before we fetch the Restaurants JSON data to display it.

At the end of main.js, we initiate a fetch() for the Restaurants JSON data as soon
as the document has finished loading, so we don't have to wait for the Google Maps
script to arrive first.

If the Google Maps script arrives before the Restaurants JSON however,
the call to updateRestaurants() would have fired and a second fetch for the
Restaurants JSON would have been made, upon the arrival of which
the whole render-tree would have been traversed again as well.