feat: Implement IndexedDB Caching for restaurant reviews
Return reviews from local IndexedDB first, to reduce time-to-display,
no matter if offline or online. Then, fetch all reviews for
a restaurant from network and replace the ones in the local IDB database
with the ones received from the network.

We replace them by removing reviews from the local database for the current
restaurant and then adding all the reviews fetched from the network for that
restaurant.

We replace them and not simply add new ones, because we could have any of the
following situations:

- Existing Reviews Deleted
- Existing Reviews Edited
- New Reviews Added.

In order to account for the deleted (non-existent) reviews,
we therefor simply clear the local database and make it in-sync with the
network database.